Replace usage of get_subscription_key() in controlling when cell blocks run.
This is because get_subscription_key() returns an error if it is neither retrievable through the environment variable or through accessing the package encryption key (when in a testthat environment). It does not return, as had been accidentally implied, an empty string "" when neither is available. The commit uses the environment variable that the function relies on directly instead. However, one should take note that if the name of the environment variable is redefined in the function, then the environment variable specified in the logic of the Rmarkdown files should be changed as well.

If the package key should be used in contexts other than that of testthat(), then the logic will need to be revisited given the complications that will introduce. For now, the simpler solution is opted for.

Also, the message shown in the vignette user_guide.Rmd has been removed from README.Rmd, as it should not appear in the output of the README.